fix(404): override @theme/NotFound so client-side hydration shows the branded page
The previous PR added `src/pages/404.tsx`, which generates the static
`404.html` Netlify serves for unknown URLs. That worked for the initial
render but on hydration the client-side router saw the unknown URL,
didn't match `/404`, and fell back to the default `@theme/NotFound`
component, replacing the branded page with Docusaurus's generic one.

Fix: extract the visual content into a shared
`src/components/NotFoundContent` and use it in both:
- `src/pages/404.tsx` (static fallback served by Netlify)
- `src/theme/NotFound/index.tsx` (client-side router fallback for any
  unmatched route)

Both code paths now render the same branded 404, so there's no
post-hydration flash.
