+++ src/theme/NotFound/index.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+import Translate, { translate } from "@docusaurus/Translate";
+import { PageMetadata } from "@docusaurus/theme-common";
+import Layout from "@theme/Layout";
+import NotFoundContent from "../../components/NotFoundContent";
+
+export default function NotFound(): React.ReactElement {
+  return (
+    <>
+      <PageMetadata
+        title={translate({
+          id: "theme.NotFound.title",
+          message: "Page Not Found",
+        })}
+      />
+      <Layout>
+        <NotFoundContent />
+      </Layout>
+    </>
+  );
+}
